requirements.txt
python-telegram-bot[webhooks]==20.7
openai==1.3.7
python-dotenv==1.0.0